+++ src/odig_api_doc.mli
@@ -5,11 +5,10 @@
   ---------------------------------------------------------------------------*)
 
 val pkg_page_mld :
-  tool:[`Odoc | `Ocamldoc ] ->
   htmldir:(Odig_pkg.t option -> Fpath.t) -> Odig_pkg.t -> string
 
 val pkg_index :
-  Odig_conf.t -> tool:[`Odoc | `Ocamldoc ] ->
+  Odig_conf.t ->
   htmldir:(Odig_pkg.t option -> Fpath.t) ->
   has_doc:Odig_pkg.t list -> no_doc:Odig_pkg.t list -> string